refactor: consolidate background job endpoints into frappe/core/api/background_jobs.py
Move the whitelisted endpoints for background tasks (get_recent_tasks,
get_cached_task_status, stop_task, retry_task), RQ jobs
(remove_failed_jobs, stop_job, get_custom_queues), the submission queue
(get_latest_submissions), scheduled jobs (execute_event,
skip_next_execution) and the scheduler (activate_scheduler,
get_scheduler_status) into frappe/core/api/background_jobs.py, marked
@frappe.public(group="Background Jobs") with full annotations and
docstrings.

Old dotted paths keep working via lazy module __getattr__ aliases;
in-repo JS callers and the Scheduled Job Type doctype actions now use
the canonical paths.
