+++ frappe/core/doctype/scheduled_job_type/scheduled_job_type.json
@@ -1,13 +1,13 @@
 {
  "actions": [
   {
-   "action": "frappe.core.doctype.scheduled_job_type.scheduled_job_type.execute_event",
+   "action": "frappe.core.api.background_jobs.execute_event",
    "action_type": "Server Action",
    "group": "Actions",
    "label": "Execute"
   },
   {
-   "action": "frappe.core.doctype.scheduled_job_type.scheduled_job_type.skip_next_execution",
+   "action": "frappe.core.api.background_jobs.skip_next_execution",
    "action_type": "Server Action",
    "group": "Actions",
    "label": "Skip Next Execution"
